--- src/main/java/org/choongang/note/services/NoteDeleteService.java
@@ -1,8 +1,0 @@
-package org.choongang.note.services;
-
-import org.springframework.stereotype.Service;
-
-// 작성한 노트 삭제 - 첨부된 파일도 함께 삭제
-@Service
-public class NoteDeleteService {
-}
