--- src/main/java/org/choongang/note/services/NoteInfoService.java
@@ -1,8 +1,0 @@
-package org.choongang.note.services;
-
-import org.springframework.stereotype.Service;
-
-// 작성한 노트 조회 - 한개 조회, 목록 조회
-@Service
-public class NoteInfoService {
-}
