--- src/main/java/org/choongang/note/services/NoteSaveService.java
@@ -1,8 +1,0 @@
-package org.choongang.note.services;
-
-import org.springframework.stereotype.Service;
-
-// 노트 작성, 수정
-@Service
-public class NoteSaveService {
-}
