--- src/main/java/org/choongang/note/validators/NoteConfigValidator.java
@@ -1,19 +1,0 @@
-package org.choongang.note.validators;
-
-import org.springframework.stereotype.Component;
-import org.springframework.validation.Errors;
-import org.springframework.validation.Validator;
-
-@Component
-public class NoteConfigValidator implements Validator {
-
-    @Override
-    public boolean supports(Class<?> clazz) {
-        return false;
-    }
-
-    @Override
-    public void validate(Object target, Errors errors) {
-
-    }
-}
