+++ src/main/java/org/choongang/notedata/controllers/adminnotecontroller/RequestNoteConfig.java
@@ -1,4 +1,4 @@
-package org.choongang.note.controllers;
+package org.choongang.notedata.controllers.adminnotecontroller;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
