+++ src/main/java/org/choongang/notedata/entities/Note.java
@@ -24,16 +24,16 @@ public class Note extends BaseMemberEntity {
     private String nName;
 
     private int rowsPerPage; // 한페이지당 노트 갯수
+
     private int pageCount; // 페이징 갯수
 
     @Column(length = 10)
-    private String locationAfterWriting; // 글 작성 이후 이동할 경로(list,view)
+    private String locationAfterWriting; // 글 작성 이후 이동할 경로(list, view)
     @Column(length = 20)
     private String skin; // 스킨
 
     @Lob
     private String category; // 분류
-
     private boolean active; // 사용 여부
 
 }
