+++ src/main/java/org/choongang/notedata/validators/NoteConfigValidator.java
@@ -6,6 +6,7 @@ import org.springframework.validation.Validator;
 
 @Component
 public class NoteConfigValidator implements Validator {
+
     @Override
     public boolean supports(Class<?> clazz) {
         return false;
