+++ src/main/resources/messages/errors.properties
@@ -1,17 +1,10 @@
 Fail.file.done= 파일 업로드 완료처리 실패하였습니다.
 
 
-
-
-//노트
+#노트
 NotFound.noteData=노트를 찾을 수 없습니다.
 
 
-NotFound.board=게시판을 찾을 수 없습니다.
-NotFound.comment=댓글을 찾을 수 없습니다.
-RequiredCheck.guestPw=비밀번호 확인이 필요합니다.
-Mismatch.password=비밀번호가 일치하지 않습니다.
-
+#노트 관리자
 NotFound.note=노트를 찾을 수 없습니다.
-
 Select.note=게시판을 선택하세요.
